Refactored routes into a routes folder.
