+++ apps/newsletters-api/src/main.ts
@@ -1,49 +1,11 @@
 import Fastify from 'fastify';
-import {
-	isNewsletter,
-	newslettersDataClient,
-} from '@newsletters-nx/newsletters-data-client';
-import newslettersData from '../static/newsletters.prod.json';
-import { makeErrorResponse, makeSuccessResponse } from './app/responses';
+import { registerHealthRoute } from './app/routes/health';
+import { registerNewsletterRoutes } from './app/routes/newsletters';
 
 const app = Fastify();
 
-app.get('/health', async () => {
-	return Promise.resolve({
-		message: 'Hello API',
-		stringFromLib: newslettersDataClient(),
-	});
-});
-
-// not using the makeSuccess function on this route as
-// we are emulating the response of the legacy API
-app.get('/newsletters', async (req, res) => {
-	const newsletters = newslettersData.filter(isNewsletter);
-	return newsletters;
-});
-
-app.get('/v1/newsletters', async (req, res) => {
-	const newsletters = newslettersData.filter(isNewsletter);
-	return makeSuccessResponse(newsletters);
-});
-
-app.get<{ Params: { newsletterId: string } }>(
-	'/v1/newsletters/:newsletterId',
-	async (req, res) => {
-		const { newsletterId } = req.params;
-		const newsletter = newslettersData
-			.filter(isNewsletter)
-			.find((newsletter) => newsletter.identityName === newsletterId);
-
-		if (!newsletter) {
-			return res
-				.status(404)
-				.send(makeErrorResponse(`no match for id ${newsletterId}`));
-		}
-
-		return makeSuccessResponse(newsletter);
-	},
-);
+registerHealthRoute(app);
+registerNewsletterRoutes(app);
 
 const start = async () => {
 	try {
